+++ CHANGELOG.md
@@ -38,6 +38,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add `calcStr` to compute textual expressions and check the output.
 - Add easings with names defined as an expression starting with '=='
 - Add extra exception handling server-agent-puppeteer to prevent freezing.
+- Add `story.getStoryToTeleportAsConfig` to provide an object access to story to teleport.
 
 ### Fixed
 - Fix `ChartLegend` in `chart-tasks` that had required typing `mark` parameter.